:sparkles: Feat: 결제 요청 API 연동

+++ app/src/main/java/com/ite/sws/domain/payment/data/PaymentItemDto.kt
@@ -14,5 +14,5 @@ package com.ite.sws.domain.payment.data
  */
 data class PaymentItemDto(
     val productId: Long,
-    val quantity: Long
+    val quantity: Int
 )
